Formatting, asserts, UNUSED, split program
The program has been split up into header and interpreter, ready so it
can be turned into a library. The code has also been reformatted (spaces
to tabs), and there is more to do on that front. Some asserts have been
added and an UNUSED macro defined to suppress needless warnings.