Fixed memory leaks, rejigged some code, more asserts
* Some minor memory leaks have been fixed
* Functions and some data structures have been renamed, changing the
case used for exported functions and structures. Some structures such as
picolVar and picolCmd have been made to be opaque. Code has been
reformatted and replaced with equivalent code (and functions) that I
find easier to read.
* More, basic, asserts have been added. These just test for NULL
pointers, and initialization is done in the correct order. More
properties; pre, post conditions, and invariants, should be checked.