Assertions, tests, code style, docs, ...
There are no new features in this commit, just more assertions about the
behavior of the system. (Once I understand the parser and evaluator
more, more assertions should be placed there).

A fair number of unit tests have been added in 'pickle.c', which caught
a bug in the unescape routines relating to escape strings in the format
'\xFF'.

There are many style changes as well, places spaces in the right place,
making more things 'const' where possible, moving code around.

A few macros have been moved and made; 'implies' for checking
material implication, and verify, which is like assert but should stay
in whether or not NDEBUG is defined. I have not found a use for
BUILD_BUG_ON in this project yet, but it is always a useful macro to
have around. A 'DEBUGGING' macro, which is '1' when NDEBUG is not
defined, and '0' when it is, can be used to optionally compile out code
without the ugly '#ifdef's that usually about, by writing code like:

   if(DEBUGGING) {
     ...
   }

With the compiler optimizing the code out.

The shell application, 'shell.tcl', has been improved by adding in
execution of a startup script in the users homepath.

Slightly more functions are now documented.