Changed API, style, code analysis, comments, ...
The major change in this commit was changing the API so all of the data
structures (apart from the optional allocator interface) were made to be
internal to the 'pickle.c' file. It is now less possible to misuse the
library. This required the addition of a few new functions to get some
more information out of the interpreter and the rearrangement of some
code so the functionality in 'main.c' stayed the same.

Some code analysis tools (clang-tidy, cppcheck) highlighted some
possible errors. They have been fixed.

There have been some minor style changes, formatting and removing
trailing white-space.

The API presented in pickle is also far more regular, it does the same
thing in all circumstances, returns an error code and passes back values
via pointers to them in its arguments.