Added memory debug option, fixed block bug
A bug in the block allocator discovered by the Fuzzer 'American Fuzzy
Lop' has been fixed, which related to its implementation of 'realloc'.
An incorrect assumption that the reallocated block was larger than the
block to be freed caused memory corruption when a block was copied from
one location to another.

Turning on memory tracing in the custom block allocator can be done with
a command line option now.