Removed some limitations, improved documentation
Some limitations on the interpreter have been moved or eased. There is
still more work to do on the internals if all restrictions are to be
lifted. The restrictions are mostly to do with maximum file size, the
biggest culprit would be the 'string' command, and the 'unescape'
function.

The documentation has been improved somewhat as well, with a fairly
decent manual page, 'pickle.1', being added. Other changes include some
changes to comments.