Turned 'argv' into a const list
Whilst useful to indicate that the 'argv' list should be 'const' it
seems to add too much noise to the function prototypes. I think I will
add the change in, then revert that change so I have a record of the
action on the mainline.

Other changes include:
* Some updates to the manual page.
* Making more functions 'inline' so they are eliminated when they have
no referents.
* Changed 'len' to 'len - 1' on some strncpy calls.
* Add some new calls in the API.