incr, gensym, documentation, getopt, limits, unknown, & more!
There is quite a lot in this commit.

There are some new commands, mostly as extensions in 'main.c' such as:

* incr (greatly increases efficiency)
* gensym
* fremove
* frename
* id
* stdin (a file routine for standard input)
* stdout (a file routine for standard output)
* stderr (a file routine for standard error output)
* for (part of the core system)
* unknown (partially implemented in the core)
* string subcommands 'dec2base' and 'base2dec' for arbitrary base
(arbitrary being bases 2 to 36 inclusive) have been added.
* '-error', '-flush' and '-eof' subcommands have been added to the file routines.

And parts of the commands have been documented, there is more to do on
this.

The arbitrary string limits on the interpreter have been removed, or
should have been, some may linger and of course they can be enabled by
compile time option (which should probably be removed).

Other changes include:

* Removal of 'pickle_getopt' from the API, and PREPACK and POSTPACK,
which simplifies the API greatly.
* A few new tests have been added.
* The documentation improved.
* The shell script command has been updated to reflect changes in the
way I/O works.
* More comments and TODOs have been added.
* Some critical bugs have been fixed.
* The run time no longer depends on 'snprintf', only 'vsnprintf' now.
* The limits of the 'long' type can be queried with the 'info' command.
* Some of the internals have been rewritten so things are not allocated
if they do not need to be, fewer copies of objects should be made. This
is due to 'picolStackOrHeap' and 'picolForceResult' being used more
often.