Removed unneeded comments, renamed commands, test linsert
* A test 'linsert' has been added. It does not quite work yet, and could
also be used for 'lreplace'. This nearly completes the list functions,
or at least the more difficult ones. They are however, incredibly
inefficient and it might be worth breaking compatibility with TCL in
order to provide something more efficient/simpler to implement for this
implementation.
* Some logical operations have been renamed. The arithmetic operators
should also probably be renamed as well.
* A lot of text has been moved from 'pickle.1' to 'readme.md'.