Updated TODO comments
A new TODO comment has been added, I think the syntax could be extended
to support the normal shell syntax for pipes and I/O redirection.
Whilst the program could be extended to do this, it does not need to be
handled within it. The idea may not be feasible, but if it was then it
could make the shell usable as '/bin/sh' replacement.