New slurp, fixed allocator, less redundant
Two bugs have been fixed, one in the 'pickle_allocate' function (which
meant the allocated memory was never actually returned to the caller,
this function had not actually been used apparently...perhaps it should
be deleted) and another in the way the new 'slurp' function was
implemented which meant it could not deal with files that you could not
call seek on.

A few redundancies and duplicated code has been removed. Things are
slightly smaller.