Removed 'variadic', added 'args' special case
The 'variadic' function has been removed and a special case has been
added whereby if the last argument is called 'args' in a procedure then
any remaining arguments in a call to that procedure are concatenated
together and passed in as 'args'. This special case is how TCL handles
things. It actually results in slightly less code, but I feel it is not
as clean, for the sake of compatibility (and because there is less
code), this feature from TCL has been added in. There is no reason to be
incompatible where possible.