clock command, info, readme,
* The 'clock' command has been added back into 'main.c' as it is useful
for performance tests, although it is not compatible with the TCL
version, so may need to be changed.
* A 'small' target has been added to the 'makefile' that produces a
smaller version of the pickle interpreter.
* More sub-commands have been added to the 'info' command and it acts
more like the TCL version of 'info' without increasing the size of the
interpreter. The weird way indexing into the hash table to inspect
procedures has been replaced with the TCL way of doing things.
* Some dead code has been removed in the form of 'ch' member in one of
the structures, which was used in an attempt to keep track of line
number (which did not work as intended, so was removed).
* A 'fatal' flag has been added, that once triggers prevents future runs
on the interpreter.
* The parser picks up on some errors it did not do before, which is
useful for the new 'info complete' subcommand.
* The performance option to the 'shell' program has been added back in.
* Some 'shell' functions have been rewritten for the new 'info'
interface.
* The 'rc' file option has been added back in.
* There is more commentary in the 'readme.md' file about design options.
* There are probably some more things that have been missed. There is a
fair amount in this commit, although the size of the core interpreter
has not actually increased that (or at all).

Other notes:

- It might be worth changing 'puts' and 'gets' so they accept more
options, 'puts' could also have an option for writing to the channel ID
'stdout' and 'stderr'.
- The interpreter is slower than it needs to be and it seems there have
been some performance regressions between this version and versions
'v1.0.1', which has can be shown when the performance programs are run
for each interpreter.
- The API really could be simplified, there are too many set/get
functions, this is also mentioned in the 'readme.md' file.
- It might be worth actually implementing the performance improvements
mentioned in the