v4 branch: profiling, list math ops, make, TODO
* Various TODO comments have been added with what I would like to
accomplish in the version 4 release
* The version number has been bumped to 4.0.0.
* A makefile target for profiling has been added, it is hoped that this
will help in speeding up the interpreter. It appears to be spending most
of its timing either allocating or fetching variables and functions.
None of this is particularly surprising.
* The mathematical operations (such as '+', '-') can now act on lists
and behave as you would expect them to. So for example a list of numbers
can be added together with plus, or a list of numbers taken away from
the first element in the list. The equality operators likewise compare
against the first element in the list.
* The number of error strings in the program has been reduced in an
effort to shrink the size of the resulting binary a little. It should be
possible to improve the error reporting without increasing the size of
anything as well.
* Some code has been refactored in the test cases, but should behave the
same.
* Comments about online (online here means within the interpreter)
help have been added.
* Some other makefile targets have been added for a 'fast' and a 'small'
version of the interpreter. There is also a 'micro' target which
as much functionality as possible.