Bug fixes, API reductions, Help string, ...
* The 'clock' command in 'main.c' has been improved and now acts more
like the TCL version of 'clock'. The code in 'shell' has been improved
to reflect this.
* A few more tests have been added to 'shell'.
* Some missing command descriptions in the 'readme.md' file have been
added.
* Some more unneeded C functions in the API have been removed.
* Version information has now been added to the modified executables
that can be built, such as 'small', and 'fast' makefile targets.
* Help strings can optionally be compiled into the program to give a
limited form of online help if an error is encountered (but unfortunately
cannot be looked up without triggering an error yet). By default the
help strings are compiled out of the program so as not to make it too
large.
* A bug in 'concatenate' has been fixed which could cause a nasty memory
corruption issue. This was found because a bug in the parser triggered
it. Not enough memory was being allocated for a string, a memory leak
was also found and fixed in the same function. The bug in the parser was
related to parsing expressions like:

	proc x {m l n} {
		puts "$m $l $n"
	}

If there was any white space after the final '}' then an extra argument
would be added to the command to be executed, causing an incorrect
error. This has been fixed in the 'picolParseSep' function by only
treating the space and tab characters as separating white-space.
* Variables can now be referred to like '${var}' instead of just by
'$var', which is useful when composing strings. This required a minor
parser change.
* More subcommands have been added to 'info', enough so a stack-dump of
the currently executing function can be made, which is demonstrated in a
new function in 'shell'.
* The version number has been bumped to version '4.0.2', which is still
a release candidate for the new version 4.
* A comment about adding the '{*}' syntax has been added, which is
sorely lacking.