Fixes for problems revealed by fuzzing (on going)
The fixes in this commit are due to problems found by the fuzzer
'American Fuzzy Lop', I have used this tool before as it has found
problems with this program before, I think the tests case I am
using are better than the ones I used before however.

There are still a number of problems that have not been fixed by
this commit, but this fixes a great many of the problems that were
found in fuzzing.

They are mostly related to the list series of functions (which really
need rewriting, they are far too complicated and could be done better).

* An assert has been added which catches an out of bounds condition. The
cause of the condition has not been solved.
* The regular expression engine failed to set a pointer when the
operator '^' was used. This causes an assertion to trigger.
* In a list operator function empty lists were not being handled
correctly.

There are still some assertions being triggered, and memory leaks on
error conditions, as well as a few more nasty bugs to be found.