Fixes for problems found by fuzzing (done)
For the test cases found by fuzzing the interpreter for a few days fixes
have been found. Another round will commence soon after this. The fixes
are similar to bugs found in the previous commits; off by one errors,
missing checks, memory leaks and a problem not setting a string pointer
in the regular expression engine caused by a exit condition failing to
do so.