Tracing, test cases and compliance
* The 'color' commands have been made to be more efficient, although
they will not respect the environment variable COLOR if it changes
during execution.
* Many more test cases have been added, they have come from a similar
Picol derivative interpreter available at
<https://github.com/dbohdan/picol>, the projects have different goals
however.
* Adding the test cases have revealed unneeded differences between TCL
and this interpreter. Some of these have been fixed in this commit,
others are still problems. As such the test cases do not all pass in
this commit, there is still work to do.
* The escape character for 'match' has been changed from '%' to '\',
inline with TCL.
* More characters trigger escaping of strings where it is needed.
* A serious 'bug' in the C implementation of the 'gets' command meant
that the format string passed to a 'printf' like function was
controllable directly from the interpreter.
* Tracing functionality has been added to the interpreter, although not
TCL compliant tracing functionality.