Fixed lreplace behaviour, heap, eval, ...
* A bug in 'lreplace' has been fixed which caused its arguments to be
treated as a non-list.
* Recursing too much with eval could cause a stack overflow, this has
been fixed.
* A 'heap' command has been added (added back in, in fact, although this
one is *much* simpler) for querying heap statistics. There are currently
far too many memory allocations taking place, even if the working set is
small. A performance bug in 'slurp' has been caught due to this addition
of this command.
* There are now case insensitive options on 'lsearch' and
'string match'.
* More test cases have been added in.
* In the 'shell' file: More code for functions that would normally
be built in features for most languages has been added in to shell,
code for 'switch' has been added, as well as 'lprepend', 'format'
and with the fix to 'lreplace' the option for executing a statement
from the command line has been added.