Ran GCC undefined behavior analyzer.
Created an executable and ran it with GCC's undefined behavior analyzer,
it reported a single problem when running the tests which should not
have been a problem anyway (as "-fwrapv" was is part of the build
options), the problem was fixed anyway and a note about doing the same
thing in the future added. The tests still passed, all that was required
was a cast of a signed value to an unsigned one in the right place and
there is no longer any undefined behavior (that has been detected
running with GCC "gcc (Debian 10.2.1-6) 10.2.1 20210110".

The same tests should be rerun with different compilers and undefined
behavior sanitizers.